Move callback utilities to ahcore.utils
